Updating fuctionalities + UI interface

+++ app/Http/Controllers/JiriController.php
@@ -2,31 +2,30 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\View\View;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Http\RedirectResponse;
 use App\Http\Requests\JiriStoreRequest;
 use App\Models\Jiri;
-use Illuminate\Http\RedirectResponse;
-use Illuminate\Http\Request;
 
 class JiriController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(): View
     {
-        $upcomingJiris = Jiri::where('starting_at', '>', now())
-            ->orderBy('starting_at')
-            ->get();
-        $pastJiris = Jiri::where('starting_at', '<=', now())
-            ->orderBy('starting_at', 'desc')
-            ->get();
+        $upcomingJiris = Auth::user()->upcomingJiris;
+        $pastJiris = Auth::user()->pastJiris;
+
         return view('jiri.index', compact('pastJiris', 'upcomingJiris'));
     }
 
     /**
      * Show the form for creating a new resource.
      */
-    public function create()
+    public function create(): View
     {
         return view('jiri.create');
     }
@@ -34,45 +33,62 @@ class JiriController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(JiriStoreRequest $request):RedirectResponse
+    public function store(JiriStoreRequest $request): RedirectResponse
     {
         $jiri = Jiri::create($request->validated());
+
         return to_route('jiri.show', $jiri);
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Jiri $jiri)
+    public function show(Jiri $jiri): View
     {
+        if (!Gate::allows('view', $jiri)) {
+            abort(403);
+        }
+
         return view('jiri.show', compact('jiri'));
     }
 
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Jiri $jiri)
+    public function edit(Jiri $jiri): View
     {
+        if (!Gate::allows('view', $jiri)) {
+            abort(403);
+        }
+
         return view('jiri.edit', compact('jiri'));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public
-    function update(JiriStoreRequest $request, Jiri $jiri)
+    public function update(JiriStoreRequest $request, Jiri $jiri): RedirectResponse
     {
-       $jiri->update($request->validated());
-       return to_route('jiri.show', $jiri);
+        if (!Gate::allows('update', $jiri)) {
+            abort(403);
+        }
+
+        $jiri->update($request->validated());
+
+        return to_route('jiri.show', $jiri);
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public
-    function destroy(Jiri $jiri)
+    public function destroy(Jiri $jiri)
     {
+        if (!Gate::allows('delete', $jiri)) {
+            abort(403);
+        }
+
         $jiri->delete();
+
         return to_route('jiri.index');
     }
 }
